bpf/tests: add test case for rev-snat case icmpv6 pkt too big.
[ upstream commit 867672f0973d620981fe933daef6f950a6fdc4dd ]

These are designed to test the flow of the packet through the
cil_from_netdev device through to the snat ingress code.

It is expected that the packets will be correctly remapped using
a combination of header fields from outer/inner/icmpv6-encapsulated
ctx data as well as from the snat map value, which contains the
source Pod IP.

This tests includes cases for TCP and UDP as the pmtu snat delivery
depends on L4 headers to perform snat map lookups.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>
Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>