daemon, option: move flag validation
Move the validation of the value provided with the --allow-localhost
flag to (*DaemonConfig).Populate, where all other flags are validated
too. This also makes the agent fatal earlier on an invalid value for one
of these flags and avoids doing a bunch of unnecessary setup work before
failing.

Also change --max-controller-interval to an uint flag which allows to
remove the validation for the value being smaller than 0.

Signed-off-by: Tobias Klauser <tobias@cilium.io>